refactor: split ProjectCard
