+++ web-components/src/components/cards/ProjectCard/ProjectCard.utils.ts
@@ -1,0 +1,10 @@
+export function getAbbreviation(unit: string): string {
+  switch (unit) {
+    case 'hectares':
+      return 'ha.';
+    case 'acres':
+      return 'acres';
+    default:
+      return 'ha.';
+  }
+}
